requirements.txt
llama-index==0.10.0
llama-index-llms-openai==0.1.0
llama-index-embeddings-openai==0.1.0
llama-index-vector-stores-chroma==0.1.0
llama-index-agent-openai==0.1.0
openai==1.3.0
chromadb>=0.4.22,<0.5.0
python-dotenv==1.0.0
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.0
requests==2.31.0
beautifulsoup4>=4.12.3
langchain==0.0.350 